Writing out a 2D mesh with netCDF
I think I will actually scrap a lot of this work. Instead of starting
from a 3D Exodus mesh and extracting 2D boundaries (which is what I
was intending to do), I am going to work straight from the 2D
mesh. This is going to be easier for me because I won't have to
understand how to relate the 3D mesh to the 2D mesh at the
boundary. And most importantly it will be more flexible. This means
that the user will have to provide the 2D meshes that correspond to
the 3D mesh. But that's ok since this is already done for Nalu.

I am keeping most of this as backup because maybe one day it will be
useful...