Preprocessing: Allow multiple invocation of a single task.
Allow user to call the same task (e.g., rotate_mesh) several times
during a run with different input parameters (different mesh parts
and/or rotation angles). Introduces a new
`task_type` option in each task section that is used to define the
actual implementation that is invoked. The section names and task names
in the `tasks` list can now be arbitary strings that are used to
differentiate the duplicate tasks.