BUGFIX: Fix bounding box calculations in CFDMesh and PreProcessing
Fix the bug where the bounding box calculation logic only examined the first
point within a bucket. In the old code the logic worked as long as element
blocks were being examined because the corner elements (and therefore their
nodes) always ended up as a single entry in a bucket. However, this is not true
of nodesets.